==> analysis/04_Compositional_Analysis.Rmd <==
---
title: 'Between Sample (Beta) Diversity Analysis: Composition of the Communities'
author: "Marian Schmidt"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: true
    toc_depth: 2
    toc_float: true
    theme: united
    highlight: tango
editor_options: 
  chunk_output_type: console
---

# Load packages 
```{r load-packages}
# Efficiently load packages 
pacman::p_load(phyloseq, iNEXT, ape, dendextend, patchwork, ggpubr, vegan, tidyverse, install = FALSE)

# load in functions and color preferences
source("code/functions.R")
source("code/colors_and_shapes.R")
```

# Load data
```{r load-data}
load("data/preprocessed_physeq.RData")
preprocessed_physeq

# What are the sample size distributions?
preprocessed_physeq %>%
  sample_sums() %>%
  data.frame()

# Remove sample that has fewest number of reads
# The fewest reads is 1839 and the next was 2267.
# So, I decided to use those extra 400 reads for more accurate beta div analysis
physeq_2267 <- preprocessed_physeq %>%
  prune_samples(sample_sums(.) > 2000, .)
```


# Normalize the reads 
## Scale the read counts

```{r scale_reads}
set.seed(777) 

# Scale the reads using function scale_reads from functions.R 

# Intution check of number of sequeces per sample
min(sample_sums(physeq_2267))


scaled_physeq2267 <-  
  physeq_2267 %>%
  scale_reads(round = "matround")
  

# Now check the sequencing depth! 
scaled_seqSums_df <-   
scaled_physeq2267 %>%  
otu_table(.) %>%  # Take the sum of the sample abundance   
colSums(.) %>%  
data.frame() %>%  
rownames_to_column(., var = "names") %>% 
 make_MA_metadata()
 
 # Rename the 2nd column 
 colnames(scaled_seqSums_df)[2] <- "TotalSeqs"   
 
 # Plot the histogram
 scaled_seqSums_df %>%  
 ggplot(aes(x=TotalSeqs)) + #fill = fraction   
 ggtitle("Histogram of Raw Read Counts Samples") +  
 geom_histogram(bins = 50) + 
 theme(legend.position = c(0.15, 0.85)) +   
 scale_y_continuous(expand = c(0,0)) +   
 scale_x_continuous(limits = c(2200, 2300)) # PLAY AROUND WITH NUMBERS! 
```

# PCoA
```{r}
# Calculate Bray-Curtis Dissimilarity 
scaled_BC_pcoa <- 
  ordinate(
    physeq = scaled_physeq2267,
    method = "PCoA",
    distance = "bray", 
    binary = FALSE
  )

# Plot the PCoA 
pcoa_fraction <- plot_ordination(
  physeq = scaled_physeq2267,
  ordination = scaled_BC_pcoa,
  color = "fraction",
  shape = "fraction", 
  title = "Scaled Bray-Curtis PCoA") + 
  geom_point(size = 5, alpha = 0.5, aes(color = fraction)) + 
  scale_color_manual(values = fraction_colors); pcoa_fraction


# Station 
pcoa_station <- plot_ordination(
  physeq = scaled_physeq2267,
  ordination = scaled_BC_pcoa,
  color = "station",
  shape = "station", 
  title = "Scaled Bray-Curtis PCoA") + 
  geom_point(size = 5, alpha = 0.5, aes(color = station)) + 
  scale_color_manual(values = station_colors); pcoa_station
```


# NMDS 
```{r NMDS}
set.seed(777)

# Calculate Bray-Curtis Dissimilarity 
scaled_BC_nmds <- 
  ordinate(
    physeq = scaled_physeq2267,
    method = "NMDS",
    distance = "bray", 
    binary = FALSE)

# NMDS by fraction 
nmds_fraction <- 
  plot_ordination(
  physeq = scaled_physeq2267,
  ordination = scaled_BC_nmds,
  color = "fraction",
  shape = "fraction", 
  title = "Scaled Bray-Curtis NMDS") + 
  geom_point(size = 5, alpha = 0.5, aes(color = fraction)) + 
  scale_color_manual(values = fraction_colors); nmds_fraction

# NMDS by station
nmds_station <- plot_ordination(
  physeq = scaled_physeq2267,
  ordination = scaled_BC_nmds,
  color = "station",
  shape = "station", 
  title = "Scaled Bray-Curtis NMDS") + 
  geom_point(size = 5, alpha = 0.5, aes(color = station)) + 
  scale_color_manual(values = station_colors); nmds_station
```


## PCoA vs NMDS
```{r PCoA-vs-NMDS, fig.width=9, fig.height=4}
# Draw both ordination methods for fraction 
## First remove legend for prettier plots 
pcoa_fraction_noleg <- 
  pcoa_fraction + theme(legend.position = "none")
# Drow side-by-side plots for fraction
pcoa_fraction_noleg+ nmds_fraction

## First remove legend for prettier plots 
pcoa_station_noleg <- 
  pcoa_station + theme(legend.position = "none")
# Drow side-by-side plots for station
pcoa_station_noleg + nmds_station
```


# Stats! 
# Check for differences in the communities using PERMANOVA 
```{r permanova}
# Calculate bray curtis for PERMANOVA
scaled_bray <- phyloseq::distance(scaled_physeq2267, method = "bray", binary = FALSE)

# pull out metadata 
metadata <- scaled_physeq2267 %>%
  sample_data() %>%
  data.frame()

# Permutational Multivariate Analysis of Variance Using Distance Matrices
# aka PERMANOVA using the adonis2 function from vegan 
# Test the hypothesis that the fraction centriods are different 
# using the bray curtis 
# Testing if the centroids of the fractions are different? 
adonis2(scaled_bray ~ fraction, data = metadata)

# Are the centroids different for stations based on bray curtis?
adonis2(scaled_bray ~ station, data = metadata)

adonis2(scaled_bray ~ date, data = metadata)

# Are they independent of each other?
adonis2(scaled_bray ~ date +  station + fraction, data = metadata)

# Is there an interaction between the variables? 
adonis2(scaled_bray ~ station * date * fraction, data = metadata)
```

# Homogeniety of dispersion test: Beta dispr
```{r}
# Dispersions between stations 
betadispr_station <- betadisper(scaled_bray, metadata$station)
permutest(betadispr_station)
```


# Cluster Analysis 

```{r hclust, fig.height= 6, fig.width=12}
# Create data for colors 
color_dat <- data.frame(sample_data(scaled_physeq2267)) %>%
  mutate(station_color = ifelse(station == "Shipping Channel", "dodgerblue4", 
                                ifelse(station == "Aransas Bay", "dodgerblue2",
                                       ifelse(station == "Copano West", "#D9CC3C",
                                              ifelse(station == "Copano East", "#A0E0BA",
                                                     ifelse(station == "Mesquite Bay", "#00ADA7",
                                                            "grey"))))),
         fraction_color = ifelse(fraction == "Particle", "firebrick3",
                                 ifelse(fraction == "Free", "goldenrod1", 
                                        ifelse(fraction == "Whole", "darkorange2",
                                               "grey")))) 

# Calculate the bray curtis hierarchical clustering
relabund_bray <- vegdist(t(otu_table(scaled_physeq2267)), method = "bray", binary = FALSE)
hclust_relabund_bray <- hclust(relabund_bray, method = "average")

# Fraction
dend_bray_frac <- as.dendrogram(hclust_relabund_bray, hang=0.1)
labels_colors(dend_bray_frac) <- as.character(color_dat$fraction_color[order.dendrogram(dend_bray_frac)])

dend_bray_station <- as.dendrogram(hclust_relabund_bray, hang=0.1)
labels_colors(dend_bray_station) <- as.character(color_dat$station_color[order.dendrogram(dend_bray_station)])

# Plot together
par(mfrow=c(2,1),
    oma = c(0,0,0,0) + 0.1, # outer margins: bottom, left, top, right 
    mar = c(5,4,0,0) + 0.1) # plot margins: bottom, left, top, right 
plot(dend_bray_frac, ylab="Bray-Curtis")
plot(dend_bray_station, ylab="Bray-Curtis")
```


# Compositional Analysis 
```{r}
phylum_df <- 
  scaled_physeq2267 %>%
  tax_glom(taxrank = "Phylum") %>%                     # Agglomerate/merge counts by phylum 
  transform_sample_counts(function(x) {x/sum(x)}) %>%  # Transform to relative abundances
  psmelt() %>%                                         # melt into long format data frame 
  dplyr::filter(Abundance > 0.01)                   # filter out phyla that are > 1% abundant 
  
# Plot the phylum distributions 
phylum_df %>%
  dplyr::filter(depth == "S") %>%
  dplyr::filter(Phylum == "Verrucomicrobiota") %>%
  ggplot(aes(x = fraction, y = Abundance, fill = Phylum)) + 
  geom_bar(stat = "identity", color = "black") + 
  facet_grid(date~station) + 
  scale_fill_manual(values = phylum_colors) + 
  theme(axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1))

```
